itsm_backend/requirements.txt
# Django ITSM Backend

Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mssql-django>=1.3
bcrypt>=4.1,<5.0
drf-spectacular>=0.27,<1.0
django-filter>=23.5,<24.0
pyodbc>=5.0
python-dateutil>=2.8
django-ratelimit>=4.1,<5.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
